Fix shortcut abort propagation to preserve pre-refactor flow
